git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@5094 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ protocolgame.cpp
@@ -1780,7 +1780,7 @@ void ProtocolGame::sendReLoginWindow(uint8_t pvpPercent)
 	{
 		TRACK_MESSAGE(msg);
 		msg->put<char>(0x28);
-		msg->put<char>(pvpPercent);
+		msg->put<char>((int16_t)pvpPercent);
 	}
 }
